requirements.txt
anyio==3.4.0
asgiref==3.4.1
autopep8==1.6.0
beautifulsoup4==4.10.0
certifi==2021.10.8
charset-normalizer==2.0.9
click==8.0.3
colorama==0.4.4
dnspython==2.1.0
email-validator==1.1.3
fastapi==0.70.0
h11==0.12.0
html5lib==1.1
httptools==0.2.0
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.3
lxml==4.6.4
MarkupSafe==2.0.1
orjson==3.6.5
pycodestyle==2.8.0
pydantic==1.8.2
python-dotenv==0.19.2
python-multipart==0.0.5
PyYAML==5.4
requests==2.26.0
six==1.16.0
sniffio==1.2.0
soupsieve==2.3.1
starlette==0.16.0
toml==0.10.2
typing_extensions==4.0.1
ujson==4.3.0
urllib3==1.26.7
uvicorn==0.15.0
watchgod==0.7
webencodings==0.5.1
websockets==10.1
